+++ Sources/MapboxNavigation/NavigationViewDelegate.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol NavigationViewDelegate: NavigationMapViewDelegate, InstructionsBannerViewDelegate {
+    
+    func navigationView(_ view: NavigationView, didTapCancelButton: CancelButton)
+}